+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-28 15:38:11
-// Commit: fix(network): fix memory leak by cancelling network subscription
+// Updated at: 2026-05-29 13:07:07
+// Commit: test(network): implement widget integration tests for network
 
 class networkComponentState {
-  final String commitRef = '1549';
-  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
-  final DateTime timestamp = DateTime.parse('2026-05-28T15:38:11+05:00');
+  final String commitRef = '1576';
+  final String lastMessage = 'test(network): implement widget integration tests for network';
+  final DateTime timestamp = DateTime.parse('2026-05-29T13:07:07+05:00');
 }